fixed many MSVC warnings
git-svn-id: file://localhost/home/tolik/svnrepo/trunk@2004 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/3rdparty/flann/constants.h
@@ -32,7 +32,7 @@
 #define CONSTANTS_H
 
 
-const int FLANN_VERSION = 1.20;
+const double FLANN_VERSION = 1.20;
 
 /* Nearest neighbor index algorithms */
 enum flann_algorithm_t {
